+++ CHANGELOG.md
@@ -4,6 +4,8 @@
   - New
     - New CLI flag `-maxtime-job` to set max. execution time per job.
     - Changed behaviour of `-maxtime`, can now be used for entire process.
+    - A new flag `-ignore-body` so ffuf does not fetch the response content. Default value=false.
+
   - Changed
 
 - v1.0.2